Fix digit input issue in script.js
Digits inputted by the user needed to be stored in variables in order
to be able to operate on them. By tracking the number of times the
digit buttons are clicked, one can assign the first digit to one
variable (when the digit button is clicked once), and the second to
another (when the digit button is clicked twice).

With this, we can now successfully call the operator functions with
the numbers inputted as arguments. However, this opens up to a separate
issue: the user can only input single digits.