Fix repeated clicks/reassignment in script.js
Previously, user could not repeatedly press the equal button after
running an operation. If the user inputted 1 + 3 then pressed equal,
display would be 4. But, if the user clicked equal again, display
would be 1, since firstNumber would be equal to 1 instead of four.

Likewise, the user could not use the signs or percentage button after
clicking equal, it would only change the original number inputted for
firstNumber instead of the result.

Decimal cannot be used yet. If decimal is clicked with the result on
display, display is '0.'.